requirements.txt
python-dotenv
pandas 
pymongo[srv]
scikit-learn
matplotlib
scipy
PyYAML
dill
-e .